#5 refactor: MainPage에서 DetailPage 개발을 하고 있어서 DetailPage로 이름 변경하고 MainPage 새로 추가
