+++ src/routes.js
@@ -1,4 +1,5 @@
 import { BrowserRouter, Routes, Route } from 'react-router-dom';
+import DetailPage from './Pages/DetailPage/DetailPage';
 import MainPage from './Pages/MainPage/MainPage';
 
 function App() {
@@ -6,6 +7,7 @@ function App() {
     <BrowserRouter basename="/pages">
       <Routes>
         <Route path="/" element={<MainPage />} />
+        <Route path="/detail" element={<DetailPage />} />
       </Routes>
     </BrowserRouter>
   );
